+++ src/lib/logger.ts
@@ -1,4 +1,0 @@
-export interface Logger {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  error(...args: any[]): void;
-}