+++ src/lib/shell-settings-resolver.ts
@@ -1,6 +1,17 @@
 import { EXTENSION_NAME } from "./const";
-import resolveOsKind from "./resolve-os-kind";
 import { Workspace } from "./adapters/workspace";
+import { ObjectMap } from "./types";
+
+const OS_KIND = {
+  darwin: "osx",
+  linux: "linux",
+  win32: "windows",
+} as ObjectMap<string>;
+const DEFAULT_OS_KIND = OS_KIND.linux;
+
+function resolveOsKind(platform: string) {
+  return OS_KIND[platform] || DEFAULT_OS_KIND;
+}
 
 export class ShellSettingsResolver {
   constructor(