+++ src/lib/types/node.ts
@@ -1,5 +1,0 @@
-import { ObjectMap } from "./collection";
-
-export interface EnvVarWrap {
-  env: ObjectMap<string | undefined>;
-}